Initial commit
I got the oauth working end-to-end, but there are some hacks in there.
Now that I have the basic flow working, I am going to start moving the
http handling into a service. I also need to investigate what is the best
way to persist all of the auth data (request-token, access-token, consumer-key).

My current assumption is that it is fine to shove the access-token into the DB once
it is obtained, but I have never actually dealt with oauth so I need to verify.