Add JSON parsing for Pocket Responses
I added GSON to parse the response for me. I got this working correctly, though
it seems like Pocket doesn't really have any way to see a fully qualified JSON response.

That should be fine since I really only care about the summary, the title, and the item id.

The current thinking is that I will show the title, summary, and a picture. You will then be able
to swipe left to delete the item or right to open the item in pocket to read.