Cleanup sign-in flow
I simplified the sign-in flow a bit to better prepare for setting
up a proper service (though, I think oauth *needs* to live in an
Activity for the callback).

I also moved the access_key/token into the SharedPreferences which seems
like a much better way of persisting that data.