Begin adding a "Check for Duplicates" task
I added a task that checks for duplicate list entries. It seems to work and
found a couple of duplicates in my list. I also learned that the Pocket API
will return a nulled-out value for an item_id that has been removed from the
list since the last getList call. I updated the bulkInsert() in the PocketProvider
to account for this. I tested it manually and it worked, but I should probably write
some unit tests the next time I have a moment to work on this app.