Make the signInTask launch automatically if you already have a token
There are two steps where we need to talk to the Pocket servers:
1 - to get an auth token
2 - to send that token to Pocket and have it turned into an access key

We really only need user interaction for the first one. They need to
click on a button that will take them to a Pocket sign-in page in their
browser where they will need to sign in. After they sign in successfully,
we no longer need their interaction. I removed the need for the second user
button press.

Still need to add some sort of loading icon so that the user knows that
something is actually happening.