Add sliding animations for the ListFragment
This implementation is really bad. I just wanted to make sure
that I was going to be able to get the right effect from the
AdapterViewFlipper.

It looks like I will, the animations actually look pretty good. I
just need to work on the timing, they are a little slow. Right now
I have the nextView() coming in from the top. I might want it to come
in from the back.