Experiment with RecyclerView
It would be extremely nice if the new recycler view did what I am trying
to force the ViewFlipper to do.

I added a test List to see if I can get the UI behaving the way I want.

I am also probably going to play around with cardView as the child view for
the recyclerView.